+++ csvEditorHtml/browser/afterDomLoaded.ts
@@ -1,0 +1,2 @@
+/* stuff to do after dom is fully loaded but before scripts */
+executeAfterDomLoadedQueue.forEach(p => p())